review comments fixed
