+++ packages/desktop-client/src/components/Titlebar.tsx
@@ -17,7 +17,6 @@ import type { CSSProperties } from '@actual-app/components/styles';
 import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
-import { Tooltip } from '@actual-app/components/tooltip';
 import { View } from '@actual-app/components/view';
 import { listen } from '@actual-app/core/platform/client/connection';
 import { isDevelopmentEnvironment } from '@actual-app/core/shared/environment';
@@ -25,6 +24,7 @@ import * as Platform from '@actual-app/core/shared/platform';
 import { css } from '@emotion/css';
 
 import { sync } from '#app/appSlice';
+import { SharedArrayBufferWarning } from '#components/SharedArrayBufferWarning';
 import { useGlobalPref } from '#hooks/useGlobalPref';
 import { useIsTestEnv } from '#hooks/useIsTestEnv';
 import { useMetadataPref } from '#hooks/useMetadataPref';
@@ -248,56 +248,6 @@ function ServerSyncButton({ style, isMobile = false }: ServerSyncButtonProps) {
         {syncState === 'disabled' ? ` ${t('Disabled')}` : null}
       </Text>
     </Button>
-  );
-}
-
-type SharedArrayBufferWarningProps = {
-  style?: CSSProperties;
-};
-
-function SharedArrayBufferWarning({ style }: SharedArrayBufferWarningProps) {
-  const { t } = useTranslation();
-  const hasSharedArrayBuffer = typeof SharedArrayBuffer !== 'undefined';
-  const isOverrideEnabled = localStorage.getItem('SharedArrayBufferOverride');
-
-  // Only show warning if SharedArrayBuffer is not supported but user has overridden the warning
-  if (hasSharedArrayBuffer || !isOverrideEnabled) {
-    return null;
-  }
-
-  const warningMessage = t(
-    'Your environment does not support SharedArrayBuffer. You may experience data loss or degraded functionality. Click to learn more.',
-  );
-
-  const handlePress = () => {
-    window.open(
-      'https://actualbudget.org/docs/troubleshooting/shared-array-buffer',
-      '_blank',
-      'noopener,noreferrer',
-    );
-  };
-
-  return (
-    <Tooltip
-      placement="bottom start"
-      content={<Text>{warningMessage}</Text>}
-      style={{
-        ...styles.tooltip,
-        lineHeight: 1.5,
-        padding: '6px 10px',
-        width: '300px',
-      }}
-    >
-      <Button
-        variant="bare"
-        aria-label={warningMessage}
-        style={{ ...style, color: theme.warningText }}
-        onPress={handlePress}
-      >
-        <Trans>Warning</Trans>
-        <SvgAlertTriangle width={13} style={{ marginLeft: '6px' }} />
-      </Button>
-    </Tooltip>
   );
 }
